nobara_sync: avoid running updatechecker at startup
Running updatechecker takes a long time and it is currently being done
to initialize a variable that needs prior initialization only in one
function. The specific function will now run updatechecker on its own.

This will get the prompt which asks for password faster by tens of
seconds, so that users can launch the updater and instantaneously insert
their password before they can leave the updater to do its own thing,
instead of having to wait for the password prompt.